requirements.txt
git+https://github.com/facebookresearch/demucs#egg=demucs
git+https://github.com/openai/whisper.git 
fastapi
uvicorn
zhconv
aiofiles
aiohttp
langchain
websockets
python-dotenv
gradio_client
celery